+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ## [Unreleased]
 * **Security**: Do not allow logging in using the password hash.
 * **Keycloak**: Allow to use the Keycloak authentication on non-http calls.
+* **Ontimize SQLStatementHandler**: Allow searching for a pattern in columns with any data type
 ## [5.9.0] - 2023-07-31
 ### Added ✔️
 * **Multi tenant**: Add feature to set tenants on non-secured requests.